+++ src/container_load.rs
@@ -75,12 +75,20 @@ impl FieldContainer {
             Self::Decimal {
                 decimal_places,
                 rounding,
+                gt,
+                gte,
+                lt,
+                lte,
                 ..
             } => decimal::load_from_py(
                 value,
                 *decimal_places,
                 rounding.as_ref(),
                 &common.invalid_error,
+                gt.as_ref(),
+                gte.as_ref(),
+                lt.as_ref(),
+                lte.as_ref(),
             ),
             Self::Date { .. } => date::load_from_py(value, &common.invalid_error),
             Self::Time { .. } => time::load_from_py(value, &common.invalid_error),